refactor : 응집도 기준으로 폴더구조 리팩토링
