+++ src/pages/SavingsCalculatorPage/components/container/ProductsTab.tsx
@@ -1,9 +1,7 @@
 import { Suspense } from 'react';
-import { useSavingsProducts } from '@entities/savings-product/api/useSavingsProductsQuery';
-import { filterProductsByCriteria } from '@entities/savings-product/lib/filterProductsByCriteria';
-import { ProductList } from '@/widgets/product-list/ui/ProductList';
-import type { SavingsProduct } from '@/entities/savings-product/model/types';
-import type { SavingsFormData } from '@widgets/savings-form/model/schema';
+import { useFilteredProducts } from '@/pages/SavingsCalculatorPage/hooks/useFilteredProducts/useFilteredProducts';
+import { ProductList } from '@/pages/SavingsCalculatorPage/components/ui/ProductList';
+import type { SavingsProduct, SavingsFormData } from '@/pages/SavingsCalculatorPage/types';
 import { EmptyState } from '@shared/ui/EmptyState';
 
 interface ProductsTabProps {
@@ -21,11 +19,7 @@ export function ProductsTab({ formValues, selectedProduct, onProductSelect }: Pr
 }
 
 function ProductsTabContent({ formValues, selectedProduct, onProductSelect }: ProductsTabProps) {
-  const { data: products } = useSavingsProducts();
-  const filteredProducts = filterProductsByCriteria(products, {
-    monthlyAmount: formValues.monthlyAmount,
-    savingTerm: formValues.savingTerm,
-  });
+  const { filteredProducts } = useFilteredProducts(formValues);
 
   return (
     <ProductList products={filteredProducts} selectedProduct={selectedProduct} onProductSelect={onProductSelect} />
