+++ src/pages/SavingsCalculatorPage/components/ui/CalculationResult.tsx
@@ -1,5 +1,5 @@
 import { colors, ListRow, Spacing } from 'tosslib';
-import type { SavingsProduct } from '@entities/savings-product/model/types';
+import type { SavingsProduct } from '@/pages/SavingsCalculatorPage/types';
 import { format } from '@shared/utils/format';
 
 interface CalculationResultProps {
